+++ http/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2026-07-04
+
+This release contains only supporting content for `trillium-client`, exposed through the unstable
+feature
+
 ## [1.4.0] - 2026-07-02
 
 ### Added